Separate users (#3)
* Separate users in the state

Add basic functionality to separate users
in the state. TODO: Add support for reading
jwt and use that id to separate users in
the state.

* Add header token verification

Adds a function that reads the request
authorization header and validates the
token. The user-id field in the token
is used to separate the users individual
states.